requirements.txt
# This file lists the Python packages required by this project.

# Core web stack
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Core libraries
torch
numpy>=1.26.0
torch-geometric>=2.7.0
rdkit-pypi
tqdm>=4.66.0

# Scientific computing
scipy>=1.10.0

# Bioinformatics
biopython>=1.81

# Service dependencies (also see requirements-service.txt)
websockets>=12.0
aiofiles>=23.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pyyaml>=6.0
prometheus-client>=0.19.0
